Added printf functionality for hex

+++ src/arch/i386/include/heap.h
@@ -3,12 +3,12 @@
 #include <stdint.h>
 
 // TODO: Initialize with address from physical memory manager
-static uint8_t *cur_heap_position;
+static uint32_t *cur_heap_position;
 
 // Error handling is done by returning a NULL pointer, 
 //  verifying that the size was allcated correctly
 void *alloc(uint32_t size) {
-    uint8_t *addr = cur_heap_position;
+    uint32_t *addr = cur_heap_position;
     cur_heap_position += size;
     return (void *) addr;
 }
